:bug: Fix issue where AgentExtensions always used shared REstateHost
- Add shared instance concurrency test.
